+++ packages/components/inputs/async-creatable-select-input/src/async-creatable-select-input.form.story.js
@@ -7,9 +7,9 @@ import { withKnobs, boolean, number } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
 import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
 import { ErrorMessage } from '@commercetools-uikit/messages';
-import Section from '../../../../.storybook/decorators/section';
-import FormikBox from '../../../../.storybook/decorators/formik-box';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import FormikBox from '../../../../../.storybook/decorators/formik-box';
+import Readme from '../README.md';
 import AsyncCreatableSelectInput from './async-creatable-select-input';
 
 const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
