+++ packages/components/inputs/async-creatable-select-input/src/async-creatable-select-input.story.js
@@ -9,8 +9,8 @@ import {
   select,
   number,
 } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import AsyncCreatableSelectInput from './async-creatable-select-input';
 
 const colourOptions = [
