+++ packages/components/inputs/async-creatable-select-input/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './async-creatable-select-input';
+
+export { default as version } from './version';
